tools/requirements.txt
docker
six
ruamel.yaml
oyaml
schemathesis
keycloak